clean(frontend): remove demo plugin data
