+++ frontend/src/components/templates/sidebar.tsx
@@ -1,37 +1,27 @@
-import PluginListProvider from "../../providers/pluginListProvider.ts";
+import { getServerList } from "../../api/api.ts";
 import isNonEmptyArray from "../../utils/utils.ts";
 import ServerList from "../molecules/serverList.tsx";
 
 export default function SideBar(props: {
-	provider: PluginListProvider;
 	onServerSelected: (serverName: string) => void;
 }) {
 	return (
 		<div id="sidebar">
 			<h2 className="text-4xl m-5">Servers</h2>
-			{createServerListOrErrorIfEmpty(props.provider, props.onServerSelected)}
+			{createServerListOrErrorIfEmpty(props.onServerSelected)}
 		</div>
 	);
 }
 
 function createServerListOrErrorIfEmpty(
-	provider: PluginListProvider,
 	onServerSelected: (serverName: string) => void,
 ) {
-	if (PluginListProvider.isLoaded(provider)) {
-		const serverList = provider.getServerList();
-		if (isNonEmptyArray(serverList)) {
-			return createServerList(serverList, onServerSelected);
-		}
-
-		return <p className="text-2xl my-3 mx-5 text-red-500">No servers found</p>;
+	const serverList = getServerList();
+	if (isNonEmptyArray(serverList)) {
+		return createServerList(serverList, onServerSelected);
 	}
 
-	return (
-		<p className="text-xs my-3 mx-5 text-gray-700">
-			Currently loading. Please wait a moment...
-		</p>
-	);
+	return <p className="text-2xl my-3 mx-5 text-red-500">No servers found</p>;
 }
 
 function createServerList(
